+++ packages/components/src/components/feedback/spinner/spinner-utils.ts
@@ -2,7 +2,7 @@ import type { BreakpointKey, BreakpointCustomizable } from '../../../types';
 import { BREAKPOINTS, parseJSON } from '../../../utils';
 
 const SPINNER_SIZES = ['small', 'medium', 'large', 'inherit'] as const;
-type SpinnerSizeType = typeof SPINNER_SIZES[number];
+export type SpinnerSizeType = typeof SPINNER_SIZES[number];
 
 export type SpinnerSize = BreakpointCustomizable<SpinnerSizeType>;