+++ packages/components/src/styles/link-button-pure-styles.ts
@@ -2,7 +2,6 @@ import type { BreakpointCustomizable, GetStylesFunction, JssStyle, Styles } from
 import {
   addImportantToEachRule,
   addImportantToRule,
-  buildResponsiveHostStyles,
   buildResponsiveStyles,
   getFocusStyles,
   getInset,
@@ -125,7 +124,7 @@ export const getLinkButtonPureStyles = (
   const hasIcon = hasVisibleIcon(icon);
 
   return {
-    ...buildResponsiveHostStyles(hasSubline ? false : stretch, getHostStyles),
+    ':host': buildResponsiveStyles(hasSubline ? false : stretch, getHostStyles),
     root: {
       display: 'flex',
       alignItems: 'flex-start',
@@ -135,11 +134,11 @@ export const getLinkButtonPureStyles = (
       boxSizing: 'border-box',
       outline: 'transparent none',
       appearance: 'none',
-      border: 'none',
+      cursor: isDisabledOrLoading ? 'not-allowed' : 'pointer',
       textDecoration: 'none',
       textAlign: 'left',
-      background: 'transparent',
-      cursor: isDisabledOrLoading ? 'not-allowed' : 'pointer',
+      border: 'none',
+      backgroundColor: 'transparent',
       color: isDisabledOrLoading ? disabledColor : active ? activeColor : baseColor,
       transition: `${getTransition('color')}, font-size 1ms linear`, // used for transitionend event listener
       ...(!hasSlottedAnchor && getFocusStyles({ offset: 1, pseudo: '::before' })),