+++ packages/components/src/types.ts
@@ -7,8 +7,6 @@ export type { HeadlineVariant, VariantType } from './components/basic/typography
 
 export type TextWeight = 'thin' | 'regular' | 'semibold' | 'bold';
 
-export type IconSize = 'small' | 'medium' | 'large' | 'inherit';
-
 export type TextColor =
   | 'brand'
   | 'default'
@@ -27,17 +25,19 @@ export type Theme = 'light' | 'dark';
 export type ThemeExtendedElectric = Theme | 'light-electric';
 export type ThemeExtendedElectricDark = ThemeExtendedElectric | 'dark-electric';
 
-export type ButtonType = 'button' | 'submit' | 'reset';
-export type ButtonVariant = 'primary' | 'secondary' | 'tertiary';
-export type LinkVariant = ButtonVariant; // alias
+export type { IconName } from '@porsche-design-system/icons';
+export type IconSize = 'small' | 'medium' | 'large' | 'inherit';
 
+export type LinkButtonPureIconName = IconName | 'none';
+export type LinkButtonVariant = 'primary' | 'secondary' | 'tertiary';
+
+export type ButtonVariant = LinkButtonVariant;
+export type ButtonType = 'button' | 'submit' | 'reset';
+
+export type LinkVariant = LinkButtonVariant;
 export type LinkTarget = '_self' | '_blank' | '_parent' | '_top' | string;
 
 export type FormState = 'none' | 'error' | 'success';
-
-// Pagination Types
-export type NumberOfPageLinks = 5 | 7;
-export type PageChangeEvent = { page: number; previousPage: number };
 
 // BreakpointCustomizable Types
 export type { BreakpointKey, BreakpointValues, BreakpointCustomizable } from './utils/breakpoint-customizable';
@@ -46,10 +46,6 @@ export type { BreakpointKey, BreakpointValues, BreakpointCustomizable } from './
 declare global {
   const ROLLUP_REPLACE_IS_STAGING: string;
 }
-
-export type { IconName } from '@porsche-design-system/icons';
-
-export type LinkButtonPureIconName = IconName | 'none';
 
 export type AlignLabelType = 'left' | 'right';
 export type AlignLabel = BreakpointCustomizable<AlignLabelType>;